+++ assets/script/global.js
@@ -1,0 +1,14 @@
+// document.addEventListener("DOMContentLoaded", () => {
+//     const navbar = document.getElementById('header');
+//
+//     const scrollThreshold = 50; // px
+//
+//     window.addEventListener('scroll', () => {
+//         if (window.scrollY > scrollThreshold) {
+//             navbar.classList.add('bg-white', 'shadow-md');
+//         } else {
+//             navbar.classList.remove('bg-white', 'shadow-md');
+//         }
+//     });
+// });
+//